refactor - move code to internal files
