+++ scopes/workspace/workspace-config-files/writers/index.ts
@@ -1,0 +1,2 @@
+export { handleExtendingConfigFiles } from './extending-config-files';
+export { handleRealConfigFiles } from './real-config-files';
